refactor: pdf step 과정 통합
